pyproject.toml
[project]
name = "ai-calorie-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.4.1",
    "feedparser>=6.0.12",
    "google-genai>=1.62.0",
    "open-clip-torch>=3.2.0",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[job-queue]>=22.6",
]
